requirements.txt
streamlit
requests
python-dateutil
dateparser
pytz
speechrecognition
pyttsx3
google-api-python-client
google-auth
google-auth-oauthlib
